Rename `handler.ts` in Cashweb module to `index.ts`
As per title. This naming is outside convention.
